[cleanup] move plotting classes to plotter
