+++ plotter/lineplotter.ui
@@ -57,7 +57,7 @@
   <customwidget>
    <class>QCustomPlot</class>
    <extends>QWidget</extends>
-   <header>qcustomplot.h</header>
+   <header>plotter/qcustomplot.h</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
